feat: rewrite library
BREAKING CHANGE: Out with the old, in with the new. This new version is
wholly incompatible with v1. The new version simply exports a function
which given a URL and Puppeteer browser, renders the page and extracts
the HTML. This leaves the implementor free to handle the browser as
they need.